added identity service

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -19,7 +20,7 @@ namespace AuctionService.Controllers
         private readonly IMapper _mapper;
         private readonly IPublishEndpoint _publishEndpoint;
 
-        public AuctionsController(AuctionDbContext context, IMapper mapper, 
+        public AuctionsController(AuctionDbContext context, IMapper mapper,
         IPublishEndpoint publishEndpoint)
         {
             _context = context;
@@ -56,12 +57,13 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
         {
             var auction = _mapper.Map<Auction>(createAuctionDto);
             // todo: add current user as seller
-            auction.Seller = "test";
+            auction.Seller = User?.Identity?.Name ?? "Unknown";
             _context.Auctions.Add(auction);
 
             var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -73,17 +75,21 @@ namespace AuctionService.Controllers
             {
                 return BadRequest("Failed to create auction");
             }
-          
+
             return CreatedAtAction(nameof(GetAuctionById), new { id = auction.Id }, newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
             var auction = await _context.Auctions.Include(a => a.Item).FirstOrDefaultAsync(x => x.Id == id);
 
             if (auction == null) return NotFound();
-
+            if (auction.Seller != User?.Identity?.Name)
+            {
+                return Forbid();
+            }
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
             auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
@@ -101,6 +107,8 @@ namespace AuctionService.Controllers
 
             return Ok();
         }
+
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -108,9 +116,14 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound();
 
+            if(auction.Seller != User?.Identity?.Name)
+            {
+                return Forbid();
+            }
+
             _context.Auctions.Remove(auction);
 
-             await _publishEndpoint.Publish(new { Id = auction.Id.ToString() });
+            await _publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
             var result = await _context.SaveChangesAsync() > 0;
             if (!result)
             {
